requirements.txt
Flask==2.3.2
Flask-JWT-Extended==4.6.0
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.1
bcrypt==4.1.2
PyMySQL==1.1.0
pyjwt==2.7.0
